Verify that x is a valid labelled variable

Verify that x is a valid labelled variable satisfying the (minimal) specification inherent in the parameter documentation of the haven::labelled() function for haven_labelled objects.

Usage

ll_assert_labelled(x)

Arguments

x

A labelled variable

Value

Invisibly returns x if the check is successful.

See also