Return a threadbare version of a vector

A bare object is an R object that has no class attributes (see rlang::is_bare_character()). A threadbare object is an atomic object (i.e. not a list(), see is.atomic()), with no attributes at all. The function returns an error if a list is passed.

Usage

threadbare(x)

Arguments

x

A vector, possibly classed, but not a list object, to strip of all attributes.

Value

A vector with the same core values as x, but with no attributes() at all, not even names().

See also